requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
requests>=2.31.0
pandas>=2.1.0
structlog>=23.2.0
pytest>=7.4.0
uuid
email-validator>=2.1.0
gunicorn

##pip install -r requirements.txt
